Restructured the UmFile class to make use of the PpFile32 class
This change farms out 32 bit pp file specific code to the PpFile32 class
and is a step along making the UmFile a genereric wrapper for all UM
files. The PpFile basically parses the file initially returning an array
of Section32 objects which form an index for the file allowing specific
data sections to be read later by the UmFile. It is still the UmFile's
responsibility to determine whether the file is actually a 32 bit pp
file.