Added ability to read 64 bit ppFiles
To do this the UmFile class now only uses 64 bit headers. This involved
swaping all instances of 32 bit data at that level for 64 bit data. The
conversion of Section32 to Section64 objects is now performed by a
constructor rather than an operator =. Throughout ppfile.cpp and
ppfile.h there are changes to do with this conversion in terms of
function inputs and outputs, but the implicit conversion between
Section32 and Section 64 means there was little code change within the
functions.
Added the actuall code to parse a 64 bit pp file in PpFile64::open.