Fixed bug in the at function of Section32 and Section34
The function was using the m_comparator variable instead of the index
passed in. Also I removed the test for whether want indices above or
below 45 as it actually makes no difference at all to the result because
(by design) both the types used are the same size and we cast to a T in
the end anyway to avoid compiler warnings.