Fixed bug in Section32 and Section64 at()
Because we were casting as __int32 and __int64 pointers the results of
the operator[] were being cast as ints. For the real variables this
meant the bit patterns were interpretted as ints then cast as floats so
giving very odd results. Now we cast to T* to get the right
interpretation of the bit pattern from the start.