Added support for reading UM FIELDSfiles
Finished the FieldsFile class by adding all the parsing code to the Open
methods. There is both a 32 and 64 bit version of this class. Also
changed the checkValidFirstWord function of UmFileBase to
checkValidWords and accept the first and fifth word of the file. this is
because we need to examine the fifth word to distinguish FIELDSfiles
from other types of files with Fixed Length Headers.