Renamed UmFileBase and derived classes to include Parser in their names.
The base UmFileBase classes ended up just being parsers for parsing the
location of the data sections of the file. Therefore UmFileBase has been
renamed UmFileParser and the derived classes have had Parser aded to
their names to reflect this usage.
For similar reasons the open method of these classes has been renamed
parse, because they parse the file and actually do not open it.