Fixed reading data section of FIELDSfiles
The data reading for FIELDS files didn't work. There was an off by one
word error in the data location. Also the function that was being used
was for pp files and expected a suffix of the data length which didn't
exist. Also the data uses 32 big words even for 64 bit files.

Basically the data reading has been moved to the parser so each file
type can have a slightly different raw data read function.