Changed variables bigEndian to swapEndiam in ppfile.cpp and ppfile.h
All variables which contained bigEndian have been swapped. This is
because actually we only check if the endianness needs swapping not if
it is actually big. We never check if the machine on which we are
compiling is big or little endian so we only know that the endianness
needs swapping not if it is big or little.