Sorted infinite error bars
Up to now infinite error bars were simply excluded from the plot. This
was due to a plplot bug that gave some bizzare lines for this case. I'm
not sure if the bug is fixed, but I've just set infinite error bars to
be drawn out beyond the plot edge.