Added DrawableItem class and some child classes
Drawable item is supposed to represent anything that is drawn on a plot.
I have added child classes of PlotData1d, PlotData2dStructurd and a
child class of PlotData1d called LineData. as you might expect LineData
draws a line plot on the plot.

Also added a class called LineStyle which holds the style for a line and
can set the current drawing style to match the parameters held. It is
not complete yet but the interface for use is defined.