Restructured header files
The header files were getting really messy, with headers depending on
other headers which is a pain when trying to include from other apps as
their is an issue with using "" and <> includes. I have basically
created a set of internal headers that can be included externally and
include all the needed other internal headers to make the headers work.
Once tested with gcc I will restructure further to ensure the external
headers work properly too.