Sorted multi d comparison functions
Using a traits struct which creates a vector with the same number of
dimensions, but a different type I got the operators which return SBOOL
to return an arbitrary number of dimensions.
While doing so I found a bug with the VECFUNC and VECINTFUNC macros
where I had a wayward const. I guess this was icked up by the move to
VS2015 as this was tested previously in VS2012.
While I was adding the ArrayEquiv traits class I added a BaseType traits
class which will get the base type of a vector. It turned out I didn't
need it, but I left it in for future use.