Fixed problem with wgdos packing pp file read
SOme packed data includes a mask at the beginning of each line of data
which flags zeros, minima or blanks. This mask is followed by the data
for the non-masked grid boxed for the line. The maks can end on a
fraction of a byte, in which case the data starts at the next byte. This
following on to the next byte was coded incorrectly (a flag was missing
to say we had been reading the mask), so the data ended up one byte off.