Fixed error with passing colour scale to plplot
It turned out that plscmap1_range does not do what I thought it
did. This function sets the ability to use a subrange of the
colour scale (from the 0-1 range) it does not set the value of the
top and bottom of the colour scale.The value of the top and bottom
of the colour scale is set by the plot call, e.g. plimage. Removed
this call from the colourscale setup function and hence removed
the passing of the data to this function - because this was used
by the auto scaling function.