Used order only prerequisites to avoid rebuilds
By having creation of .o files dependent upon the build directory
they get rebuilt whenever the directory timestamp changes. To avoid
this add the build directory as an order only prerequisite by putting
it after a |. This means that the build directory rule will be caled
frst, but does not force a rebuild if the timestamp is newer.