Had a good mess around with unicode support
Basically I decided that the best way to deal with unicode was to use
utf32 strings and convert to either utf8, utf16 or ansi depending on the
OS and scenario. So I typedefed sci::string to be a std::u32string,
added a macro sU() which when wrapped around string litrals ensures they
are utf32 and then created function that will convert to the various
OS native types.

Ideally on windows the native type would be a wstring, however there are
some libraries that don't accept wstrings for filenames, therefore the
only option is to use ansi and hope that any special characters are
represented by the codepage. On that basis I included functions to
convert to codepage as well as unicode. Note that on linux, all these
functions result in utf8 encoded strings.

I still have to go through all the strings in the library and replace
std strings and std wstrings with sci strings.