Updated all the netcdf code to use sci strings in inputs and oututs
This give unicode support accross the board for data stored within
netcdf files. In the netcdf files the data is stored as utf8 and there
are some std strings which represent those directly written/read
strings. But the whole user interface now uses sci strings.

One issue is that the netcdf library only accepts file names as 8 bit
strings and passes them directly to the fopen style functions so this
only gives ansi support on windows.