Added code to deal with physical values with units
Basically we now have a class Physical, which is templated. The template
parameter is a class caled EncodedUnit, which is again templated with a
64 bit integer which encodes the dimensions and an 8 bit integer which
encodes the 10 power (i.e. nono, kilo etc). The Physical stores the
value itself.

Because the EncodedUnit has the dimensions encoded as a template
parameter we can do compile time checks to ensure that we are only
adding units of the same dimension.

Things seem to work pretty well. There are shortcut classes which are
the standard SI units. There is a Physical that accepts two units
combined which makes things nice and easy for creating e.g a m s-1.

All sensible operators are overloaded as it pow. Note that there are
some limitations on pow because of how things are encoded, but they
actually make physical sense too.

It may be good to use a parameter pack/variadic templates to make
combining many units easier.