Restructured Units
Moved any combination of units (multipying, dividing, raising to power)
into the encoded units classes, not the Physical class, This means there
is just one Physical class.
Units can be combined using the Multiplied.. and Divided.. EncodedUnit
classes and the multiplication and division can be repeated as many times
as needed.

The fact that this change was best was realised as I was writing code to
get the name of a unit, so this commit also includes code to sort the name
of units too.

As part of this I added compile time checks that exponents with metric
prefixes are used.