Changed the way sci::Unitless works
sci::Unitless now cannot have an exponent. I think trying to raise
a unitless to an integer power already would have failed but it
didn't work for unit string output. E.g. a metre multipled by a
unitless with an exponent of -3 would have to be a m, which loses
the milli bit or a m m, which could be confused with a m2.

Instead we jsut remove the ability for unitless to have exponents,
add a Percent, PerMille and BasisPoint class, which do have
exponents and symbols, then specialize the pow function so that
unitless to a power is still a unitless and all parts pers raised
to powers gives a Unitless.