Changed the sci::TimeInterval typedaf to be a sci::Physical<Second<>>
This broke all the time related operator overloading in time.cpp. The reason
evaded me for ages. It turned out it was as follows

The Physical operator overloading was in the sci namespace
The time operator overloading was in the global namespace
C++ considers that a function in the local namespace overrides
a function in an upper namespace even if the function has the
wrong parameter types.
This meant the code in time.cpp found the Physical overloads and
did not look further into the global namespace.
These overloads didn't match so the error was reported.

The fix for this was to put the time operator overloads in the sci
namespace too.