Made changes to unit conversion code
This basically came from wanting physicasToValues to be able to output
with a different value type to its input.

This meant creating ssome free functions that I don't want users to call,
so I put them in a namespace with the word private in. This gives us a set
of recursive template functions that output the correct type.

As part of this I had problems with the UnicConverter structs and then
realised they are actually not needed so I stripped them out.

There were some compile errors which required me to add the word template
in at various points. I think this is like using the typename keyword, but
it's not something I am that clear on.

I also needed to move the pow10 function nearer the top of the file.