Added replacement code to calcuate eigenvales/vectors
This is some code released under MIT licence and based on numerical
recipes. I had to hack the code a bit as it was written for C99 which
VS doesn't support. Also, it was just a c file, no build system. So
I figured it was easiest to just pull the file into the svector
code base and write a header for it.

The code I had to hack about with was the complex number arithmatic.
C99 has built in support for complex numbers. VS, has complex.h which
includes basically the same functionality, but the declaration style
used cannot be compiled, nor is their support for arithmetic operators
so I added some functions for arithmatic and call these instead.