Fixed and error with static assert in Units.h
It turns out that gcc expands static_asserts even in templates that are
not instantiated, but ONLY IF the test has no dependence upon the template
parameter. The standard allows, but doesn't require it to do so.

Hence we need to change ay static_assert(false, ...) calls. This commit
was just checking that the change worked for one instance.