Some modifications to units.h - should have no impact on use
Added constexpr to the Physical class

In the copy constructor I do explicit casting. Feels right even if
we are going from a Physical double to a Physical float as it
is basically a cast by the user. Don't do this for operator=.

Add a pragma to avoid a warning about not initialising Physical m_v
when an object is created by the default constructor. This is by design.

restructure te pow function to make it a bit more readable.