Changes to time code
This all started because I wanted a numeric_limits<UtcTime>

It made me realise there were a number of things that I could do
better with the UtcTime class.
First many of the functions could be constexpr. This would help
with the numeric_limits class as the methods need to be constexpr.

However the calls to gmtime and mkgmtime were not constexpr. I then
found out that they do not work for dates before 1st Jan 1970 and
they have a limit in the future around year 3000 or something.

I decided that wasn't satisfactory, so I wrote my own versions of
those functions that were constexpr and covered basically any date
for which the year fits in a signed 32 bit int (std::tm::tm_year),
which is about +/-2 billion years - not quite back to the big bang
but it will do I'm sure. Plus I made all the functions I could into
constrexpr.

Then I got loads of linker errors in the code that uses UtcTime.

It turs out that constexpr quite sensibly implies inline which
requires the function definition be available to the funtional block
at compile time. Unfortunately VS doesn't do that check. Instead
any functions it decides to inline just don't get put in the library,
meaning linker errors are generated.

So I had to move most of the functions into the header rather than
the cpp file.