Added std::numpunct specialization
std::numpunct is set up so that it is imported from the c runtime
dll for dynamic builds. But the dll only contains specializations
for char and wchar_t.

I've added a specialization for char16_t so that dynamic builds
work strings using this type.

the numpunct class basically defines the decimal point, thousand
separator and the text for true and false when converting numbers
and bools to text.