Bug change to units.h, now supporting fractional powers
I realised that the RootedUnit class was actually wrong. And it
became very clear that it was very easy to create fractional
powers that would generate compile errors. for example

RootedUnit<MultipliedUnit<Second,Second>,2>

Would fail, because it was not equivalent to

RootedUnit<PoweredUnit<Second,2>,2>

In the first example the two second units are kept separate by
design and hence it all fell apart.

Plus the inability to deal with fractional powers meant that

PoweredUnit<RootedUnit<Second>2,2>

would also fail.

So I added in numerators and denominators for the encoded powers
and also the exponent powers. This now means that we can cope
with fractional powers from 128 to 1/128 to 0 to -1/127 to -127.

I still have some testing to do to make sure I haven't added any bugs
but it all compiles and the current tests work.

Note that I haven't changed the interface to the metric units, they
are still initialised with whole powers. I'm not sure if I will
change them or not. I think it is likely that initialising fractional
powers is something that is unlikely to happen much.

I also added constexpr versions of pow and root. These will need
a bit of testing. root uses Newton Raphson.