Fixed some errors in scaled units
There was a problem with a left over reference to power not in fraction form

Also i had messed up all my imperial unit conversions and they were inverted.
It was easier to invert the angle and time values and swap * and / in the maths
than change the imperial units. Plus it seemed to make more sense.