Moved wxString functions from sstring.cpp to sstring.h
These functions depend on the wxString header being included for
them to compile/link. Previously I had a check for __WXWINDOWS__
define in both the .h and .cp file. But of course, this created
linker errors becasue my library didn't include that define.

I now check for __WXWINDOWS__ in the header file. If it is defined
I include wx/string.h and then have the functions defined as inline
in the header file. This means they are always available correctly
in wxWidgets applications and are totally excluded from non
wxWidgets applications.