Made Brent minimisation return the residual and added a gradient descent minimisation.
By returning the residual we avoid an extra call to the functor in the calling routine - this makes a performance difference where aonly a small number of calls are made to brent but whe have a large number of parameters to search.

Added a Gradient descent method which calls brent for each direction. This Should be much faster than Powell where we have gradient information.