Added a renderer class and started adapting the plot code to use it
The renderer class is intended to be a wrapper around any other renderer.
It provides a standard user interface.
It currently has code for lines, rectangles, elipses and text. Text is
stil a work in progress. Formatted text can be achieved using some basic
LaTeX style formatting options. It may be that this gets adapted in the
future to be a bit more formal. The formatted text needs the option of
rotation to be added.

The plot axis code has now been adapted to use the renderer code