Added LineData to the new plot rendering regime which ended up swapping splotaxis to using Point for its location
This was intended to be a small change to just add a new Draw method
to LineData and a method to Renderer to draw a segmented line.

However, I wanted to hae an effective way for an axis to change the
value to a Point on the page, so this became a big change.