some bug fixes and feature additions to plot and graphics
Moved the FillData class to having a new renderer style

Added the option to ignore some plots for the next render and also
add at a particular depth on the stack, remove and swap plots. This
is useful for animating or real-time modifiable plots.

Did some tidying in graphics.h and made the constructors of
GraphicsVector private, so users can only create Points and Distances