Added clipping to the graphics
This is set up so that whenever you set a clipping region, a Clipper
object is returned and when this Clipper object is destroyed the
clipping region is removed. TheClipper objects cannot be copied
created or moved outside the Renderer class so they will always be
destroyed on scope exit.

The wxRenderer struggles a bit with this as it only has the
ability to keep track of rectangular clipping regions. So when we
remove a clipping gregion we must remove all clip regions and then
reapply all except the last one.