Fixed a grid "bug"
I was getting a compiler crash. It seemed that the issue was to do
with recursive calling when indexing a transform pair grid. The
compiler perhaps couldn't work out when the recursion ended.

I think the fix was to remove some constexpr if statements and
replace them with requires statements.

I also cleaned up some unused default constructors while trying
to work out the cause.