Found a bug with mixing scalars and grids in math calculations
This was due to poor constraints/checks on the operator[] in
gridtransform_view. The function was getting the offset
from the shape of the first operand, but if the first
operand was a scalar then this didn't work. This caused an
internal compiler error. The operators[] functions have now
been laid out in a better order with comments to make it clear
what each of them do and the offending code has been fixed with
some if consexpr checks. i also took the oportunity to remove the
commented out functions that return non-const references (the
reurn values are not l-values so non-const references make no
sense).