Found some more issues with gridpairtransform_view
I still has an internal compiler error when assigning a GridData
from a gridtransform_view when one of the pair was a scalar. This
Turned out to be a bad static cast where I got the two iterator
types swapped around.

In finding this bug, I found a few other bugs where scalars were
not treated correctly, especially when they were the first of the
pair.

I think these are all now done with.