Made the Scale class templated so it can hold types other than doubles
This warranted some template based changes to the Data class as this
needs to interact with a tuple of scales.

I also made the colourscale class templated as well

All the other scale derived classes currently just use double as
the template parameter.

Note that it became clear when doing this that returning logged
vs linear data can result in different types - e.g. when using
Units the linear data could be metres, but the log data needs to be
unitless. This is going to add to code complexity unless I can think
of a way to deal with it more effectively.