Updated the rendering a bit
This is inspires somewhat by th AGG library. When rendering a shape
the code works out a scanline and writes this. However, one difference
is that for partially covered pixels, we use a bitmask which indicates
which parts of the pixel are covered and we will wait to render these
later, onece we have examined all shapes that are on this scanline.

This means that when we have multiple shapes partially covering a pixel
we can get a better approximation of the coverage and overlap by the
different shapes and this will help us avoid seams where the background
shows through tesselating shapes, or a shape behind shows through when
shapes lay exactly on top of each other.

Currently the only shape is an axis aligned rectangle and there is no
overarching class to hold shapes and render them.