Fixed a bug in the rederer related to the base
The condition for telling if a pixel was clipped by the base was
incorrect and it thought that an entirely clipped base (base on the
pixel boundary) was not clipped.

Fixed this, and also slightly restructured the intersaction of side
and base clipping. The base clipping now only needs to be checked
(and should only be checked, as clipped regions overlap) if there
is no side clipping