Made all the top level plot classes templated
This now allows them to somewhat take advantage of the templating
of the Data and lower classes.

HOWEVER, I've only tested them using doubles as the template
parameter. I expect that most classes will fail to compile if you
use a type where the logged type is not the same as the linear type.
So you sould be fine with double and float, but will fail for
sci::Physicals.

I need to make a decision about how to deal with logs. I could allow
a deeply nested if and just get each point, allowing the Data class to
choose whether it wants logged or linear data. This sounds slow, but
the reality may be that, because for a whole dataset, isLog will return
the same answer, the CPU branch predictor may make this just as fast as
a single isLog call.

I may need to do some checks and benchmarking